Replace /usr/bin/python with /usr/bin/env python
If you are running ossipee in a virtualenv (which is a good idea) then
running scripts that link directly to /usr/bin/python will bust you out
of that virtualenv and it will be unable to find the required packages.

By running with /usr/bin/env python instead it will source the python
from the virtualenv you are using.