Add setuptools tooling
Add a setup.py so that I can pip install -e . into a virtualenv
rather than have to keep jumping back to the folder to run ossipee
commands.

I use PBR because it's good for simple things like this and it makes it
easier to reuse logic from other openstack projects.