Clear both IP and hostname from known_hosts
Currently we clear the ssh key associated with an IP address from the
known_hosts file as this changes every time you redo an ossipee run. In
my case i've got a hosts entry because i need to use the hostname when
connecting via browser for SSL and testing websso flows.

This means i also tend to use the hostname when SSH-ing but this doesn't
get reset.

Reset the hostname ssh key as well, but don't try to ssh connect to it
as it may not have an entry in hosts.