Use the public key from the plan
The public key specified in the config file is being ignored in favour
of the first public key in nova. This public key is not the one I use
and so it fails to connect.

Interestingly even though ssh fails it goes through its 5 attempts to
connect and continues on without an error message even though it
couldn't actually connect to the VM.